% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlfa.R
\name{bnlfa}
\alias{bnlfa}
\title{Fit a BNLFA object.}
\usage{
bnlfa(x, response = c("nonlinear", "linear"), noise_pooling = c("lv",
  "partial-pool", "pool", "none"), prior = c("normal", "sign"),
  init = "random", model_mean_variance = TRUE, sign_bits = NULL,
  k_means = NULL, t0_means = NULL, k_sd = NULL, t0_sd = NULL,
  lambda = 1, ...)
}
\arguments{
\item{x}{Either an \code{SCESet} from \code{scater} or a
cell-by-gene (N by G) matrix of non-negative values representing gene expression.
log2(TPM + 1) is recommended.}

\item{response}{The type of factor analysis, either \code{nonlinear} (default) 
or \code{linear}}

\item{noise_pooling}{The pooling of the precision parameters, either \code{pool}, \code{partial-pool} 
(default) or \code{none}}

\item{prior}{The type of prior specification, either \code{normal} (default) or \code{sign}}

\item{model_mean_variance}{Logical. If \code{TRUE} then the variance as modelled as a function of the mean.}

\item{sign_bits}{A vector (length G) of sign bits}

\item{k_means}{G mean activation strength parameters}

\item{t0_means}{G mean activation time parameters}

\item{k_sd}{Optional standard deviations for k parameters}

\item{t0_sd}{Optional standard deviations for t0 parameters}

\item{...}{Additional arguments to \code{rstan::sampling}}
}
\value{
An object of type \code{bnlfa_fit}
}
\description{
Fit a Bayesian non-linear factor analysis model given some single-cell
gene expression data.
}
\details{
This function takes either a \code{SCESet} object or an expression matrix
and returns a \code{bnlfa_fit} object including posterior traces for all
variables.
}

