% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_consistency}
\alias{cluster_consistency}
\title{Cluster the consistency matrix}
\usage{
cluster_consistency(cmat, n_clusters = 2:9)
}
\arguments{
\item{cmat}{A consistency matrix returned by \code{consistency_matrix}}

\item{n_clusters}{The number of clusters for which to calculate the BIC}
}
\description{
Clusters the consistency matrix, using Gaussian Mixture Modelling
on its first principal component. The number of clusters is chosen
such that the BIC is maximise
}
\examples{
data(oui)
cmat <- consistency_matrix(oui)
clusters <- cluster_consistency(cmat)
}
\seealso{
consistency_matrix
}
