% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confusion_matrix_ordered}
\alias{confusion_matrix_ordered}
\title{Order confusion matrix by pseudotime}
\usage{
confusion_matrix_ordered(oui, cmat = NULL)
}
\description{
Order confusion matrix by pseudotime
}
