% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R, R/utils.R
\name{consistency}
\alias{consistency}
\alias{plot_consistency}
\alias{consistency}
\alias{consistency_matrix}
\alias{consistency}
\alias{consistency_matrix_ordered}
\title{Create, order, and plot a consistency matrix}
\usage{
plot_consistency(oui, cmo = NULL, interpolate = FALSE)

consistency_matrix(oui)

consistency_matrix_ordered(oui, cmat = NULL)
}
\arguments{
\item{oui}{A \code{ouija_fit} object}

\item{cmo}{An optional ordered consistency matrix}

\item{interpolate}{Passed to \code{geom_raster}}

\item{cmat}{A pre-computed consistency matrix returned by \code{consistency_matrix(oui)}}
}
\value{
A cell-by-cell consistency matrix (ordered by posterior pseudotime
if \code{consistency_matrix_ordered}), or a \code{ggplot2} plot object
displaying the ordered consistency matrix (if \code{plot_consistency_matrix})
}
\description{
The consistency matrix is the N-by-N (for N cells) matrix where the entry in the
i-th row and j-th column describes the empirical probability that cell i is ordered
after cell j. These functions allow the calculation, ordering (by posterior pseudotime)
and plotting of the consistency matrix. See \code{cluster_consistency} to cluster
the cells using Gaussian Mixture Modelling.
}
\details{
The consistency matrix is useful for clustering the pseudotime trajectory into discrete
stages, as if there are regions where the empirical ordering probability is around 0.5
then there is large uncertainty as to the ordering, implying the cells are of roughly
the same type. If the probability is closer to 0 or 1 then there is less uncertainty
and the cells are probably undergoing a smooth transition.
}
\examples{
data(oui)
plot_consistency(oui)
cmat <- consistency_matrix(oui)

cmo <- consistency_matrix_ordered(oui, cmat)
}
\seealso{
cluster_consistency
}
