% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\docType{data}
\name{example_gex}
\alias{example_gex}
\title{Example gene expression matrix}
\format{An object of class \code{matrix} with 400 rows and 11 columns.}
\usage{
example_gex
}
\value{
A 400-by-11 example expression matrix
}
\description{
A matrix containing some example gene expression data for 
400 cells and 11 genes, the first 9 of which exhibit switch-like
expression and the
}
\examples{
data(example_gex)
}
\keyword{datasets}
