% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{fit_rss}
\alias{fit_rss}
\title{Residual variance after fit}
\usage{
fit_rss(bm, weights = NULL)
}
\arguments{
\item{bm}{An object of class \code{bnlfa_fit}.}

\item{weights}{Cell specific weights.}
}
\description{
Residual variance after fit
}

