% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genereg}
\alias{genereg}
\alias{gene_regulation}
\title{Differences in regulation timing}
\usage{
gene_regulation(oui)
}
\arguments{
\item{oui}{An object of type \code{ouija_fit}}
}
\value{
A \code{data.frame} as described above
}
\description{
Computes a data frame with a row for every pairwise comparison
between genes, with a column for the posterior mean difference 
in regulation timings, along with the lower and upper 95% 
highest posterior density credible intervals, and a logical
indicating whether the differences in regulation timings
are significant.
}
\examples{
data(oui)
genereg <- gene_regulation(oui)

}
