% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map}
\alias{map}
\alias{map_pseudotime}
\alias{map}
\alias{switch_times}
\alias{map}
\alias{peak_times}
\alias{map}
\alias{switch_strengths}
\title{Extract the MAP estimates from a \code{ouija_fit}}
\usage{
map_pseudotime(oui)

switch_times(oui)

peak_times(oui)

switch_strengths(oui)
}
\arguments{
\item{oui}{An object of class \code{ouija_fit}.}
}
\value{
MAP estimates of the specified quantity.
}
\description{
Extract the MAP (maximum a posteriori, aka posterior mode) 
estimates of the pseudotimes, switch times, peak times
or switch strengths from a \code{ouija_fit}.
}
\examples{
data(oui)
tmap <- map_pseudotime(oui)
t0map <- switch_times(oui)
pmap <- peak_times(oui)
kmap <- switch_strengths(oui)
}
