% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlfa.R
\name{map_pseudotime.bnlfa_fit}
\alias{map_pseudotime.bnlfa_fit}
\title{Extract the MAP pseudotime values from a \code{bnlfa_fit}}
\usage{
map_pseudotime.bnlfa_fit(bm)
}
\value{
MAP pseudotime vector of length N
}
\description{
Extract the MAP pseudotime values from a \code{bnlfa_fit}
}

