% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\name{map_pseudotime.ouija_fit}
\alias{map_pseudotime.ouija_fit}
\title{Extract the MAP pseudotime values from a \code{ouija_fit}}
\usage{
\method{map_pseudotime}{ouija_fit}(bm)
}
\value{
MAP pseudotime vector of length N
}
\description{
Extract the MAP pseudotime values from a \code{ouija_fit}
}

