% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\docType{data}
\name{oui}
\alias{oui}
\title{Precomputed Ouija fit}
\format{An object of class \code{ouija_fit} of length 13.}
\usage{
oui
}
\description{
The result of calling \code{oui <- ouija(synth_gex)} to avoid
large times computing while testing.
}
\examples{
data(oui)
}
\seealso{
synth_gex
}
\keyword{datasets}
