% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\name{plot.ouija_fit}
\alias{plot.ouija_fit}
\title{Plot a \code{ouija_fit}}
\usage{
plot.ouija_fit(bm, what = c("trace", "map", "diagnostic", "dropout"), ...)
}
\arguments{
\item{bm}{An object of class \code{ouija_fit}}

\item{what}{One of
\itemize{
\item \code{trace} This produces a heatmap of gene expression as a function of pseudotime
across different pseudotime samples. Underlying call is to \code{\link{plot_ouija_fit_trace}}.
\item \code{map} This plots gene expression as a function of the MAP pseudotime with a red
line denoting a LOESS fit (showing the overall trend). Underlying call is to
\code{\link{plot_ouija_fit_map}}
\item \code{diagnostic} This returns trace and autocorrelation plots of the log-posterior
probability. Underlying call is to \code{\link{plot_ouija_fit_diagnostics}}
\item \code{dropout} Returns a plot showing the relationship between latent expression
value and dropout probability. 
Underlying call is to \code{\link{plot_ouija_fit_dropout_probability}}
}}

\item{...}{Additional arguments passed to the corresponding functions}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot a \code{ouija_fit} object. Returns either a trace fit, MAP fit or MCMC diagnostic fit.
See the individual function calls (described below) for more details.
}

