% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlfa.R
\name{plot_bnlfa_fit_diagnostics}
\alias{plot_bnlfa_fit_diagnostics}
\title{Plot MCMC diagnostics.}
\usage{
plot_bnlfa_fit_diagnostics(bm, arrange = c("vertical", "horizontal"))
}
\arguments{
\item{bm}{A \code{bnlfa_fit} object}

\item{nrow}{Number of rows. If 1, plots are side-by-side; if 2, plots are vertically aligned.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot MCMC diagnostics (traceplot and autocorrelation) for a \code{bnlfa_fit} object
}

