% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlfa.R
\name{plot_bnlfa_fit_dropout_probability}
\alias{plot_bnlfa_fit_dropout_probability}
\title{Plot dropout probability}
\usage{
plot_bnlfa_fit_dropout_probability(bm, posterior_samples = 40)
}
\arguments{
\item{bm}{An object of class \code{bnlfa_fit}}

\item{posterior_samples}{Number of posterior samples to add to the plot. If 0, only
the MAP estimate is plotted.}
}
\value{
An object of type \code{ggplot}
}
\description{
Plot the probability of a dropout as a function of latent expression value. In bnlfa
this is implemented via logit regression, so the probability of dropout is related
to latent expression \eqn{\mu_{ig}} via
\deqn{\frac{1}{1 + \exp(-(\beta_0 + \beta_1 \mu_{ig}))}}
The red curve shows the MAP estimate of the relationship, while the grey lines show 
posterior samples of the relationship.
}

