% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlfa.R
\name{plot_bnlfa_fit_trace}
\alias{plot_bnlfa_fit_trace}
\title{Plot heatmaps of gene expression changes}
\usage{
plot_bnlfa_fit_trace(bm, samples = 50, genes = 1:min(bm$G, 5),
  output = c("grid", "plotlist"), show_legend = FALSE, ...)
}
\arguments{
\item{samples}{Number of posterior pseudotime samples to use}
}
\description{
Plot heatmaps of gene expression changes
}

