% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\name{plot_ouija_fit_behaviour}
\alias{plot_ouija_fit_behaviour}
\title{Plot gene expression as a function of MAP pseudotime}
\usage{
plot_ouija_fit_behaviour(oui, genes = seq_len(min(oui$G, 6)),
  expression_units = "log2(TPM+1)")
}
\arguments{
\item{oui}{An object of class \code{ouija_fit}}

\item{genes}{A vector that subsets the gene expression matrix. Defaults to the first \code{g}
genes, where \code{g} is either 4 or the number of genes in the model if less than 4.}

\item{expression_units}{The label for the y-axis of the plot. Defaults to
"log2(TPM + 1)".}
}
\value{
An object of class \code{ggplot2}
}
\description{
Plot gene expression as a function of the MAP pseudotime with a red
line denoting a LOESS fit (showing the overall trend). Genes are plotted with
one per grid square (using a call to \code{facet_wrap(~ gene)}).
}
\examples{
data(oui)
plot_ouija_fit_behaviour(oui)
}
