% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\name{plot_ouija_fit_comparison}
\alias{plot_ouija_fit_comparison}
\title{Plot heatmaps showing comparisons of measured data and imputed}
\usage{
plot_ouija_fit_comparison(oui, return_plotlist = FALSE)
}
\arguments{
\item{oui}{An object of class \code{ouija_fit}.}

\item{return_plotlist}{If TRUE then the list of \code{ggplot}s is returned
instead of being plotted with \code{cowplot::plot_grid}}
}
\value{
Either a list of plots of class \code{ggplot} or a single 
\code{ggplot} showing them
}
\description{
Plot heatmaps showing comparisons of measured data and imputed
}
\examples{
data(synth_gex)
oui <- ouija(synth_gex, strengths = 5 * c(1, -1, 1, -1, -1, -1), iter = 100)
plot_ouija_fit_comparison(oui)
}

