% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\name{plot_ouija_fit_diagnostics}
\alias{plot_ouija_fit_diagnostics}
\title{Plot MCMC diagnostics.}
\usage{
plot_ouija_fit_diagnostics(bm, arrange = c("vertical", "horizontal"))
}
\arguments{
\item{bm}{A \code{ouija_fit} object}

\item{nrow}{Number of rows. If 1, plots are side-by-side; if 2, plots are vertically aligned.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot MCMC diagnostics (traceplot and autocorrelation) of the log-posterior probability
for a \code{ouija_fit} object.
}
\details{
Further assessment of convergence can be done using \code{rstan} functions on the
underlying STAN object (accessed through \code{bm$fit}).
}

