% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\name{plot_ouija_fit_pp}
\alias{plot_ouija_fit_pp}
\title{Prior-posterior density plots for activation parameters}
\usage{
plot_ouija_fit_pp(oui, genes = seq_len(ncol(oui$Y)), param = c("k", "t0"))
}
\arguments{
\item{oui}{An object of class \code{ouija_fit}}

\item{genes}{A numeric vector indicating indices of genes to plot}

\item{param}{Either plot the activation strength \code{k} parameter or
activation time \code{t0} parameter}
}
\value{
An object of class \code{ggplot2}
}
\description{
Prior-posterior density plots for activation parameters
}

