% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotexprs}
\alias{plotexprs}
\alias{plot_switch_times}
\alias{plotexprs}
\alias{plot_peak_times}
\alias{plotexprs}
\alias{plot_expression}
\title{Plot fitted expression, switch times, and peak times}
\usage{
plot_switch_times(oui)

plot_peak_times(oui)

plot_expression(oui, ncol = 2, nrow = NULL)
}
\arguments{
\item{oui}{A \code{ouija_fit} to plot}

\item{ncol}{The number of columns for the expression plot}

\item{nrow}{The number of rows for the expression plot}
}
\value{
A \code{ggplot2} object plotting the expression, switch times, or peak times
}
\description{
Plot the expression along pseudotime with the maximum a posteriori (MAP)
mean function as fitted by Ouija, or the MAP estimates of the
switch and peak times along with 95% HPD credible interval error bars.
}
\examples{
data(oui)
plot_expression(oui)
plot_switch_times(oui)
\dontrun{
plot_peak_times(oui)
}

}
