% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predexprs}
\alias{predexprs}
\alias{predicted_expression}
\alias{sample_predicted_expression}
\title{Predicted expression (mean and sampling)}
\usage{
predicted_expression(oui)

sample_predicted_expression(oui)
}
\arguments{
\item{oui}{An object of class \code{ouija_fit}.}
}
\value{
A matrix of the same dimension as \code{oui$Y} containing 
the MAP mean expression or a posterior sample of the mean expression.
}
\description{
Extract the posterior mean of the predicted expression (ie the sigmoid
or transient function), or a sample from the posterior.
}
\examples{
data(oui)
pexp <- predicted_expression(oui)
sample_pexp <- sample_predicted_expression(oui)
}
