% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ouija_fit}
\alias{print.ouija_fit}
\title{Print a \code{ouija_fit}}
\usage{
\method{print}{ouija_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ouija_fit}.}

\item{...}{Additional arguments.}
}
\value{
A character string representation of \code{x}
}
\description{
Print a \code{ouija_fit}
}
\examples{
data(oui)
print(oui)
}
