% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtests.R
\name{reg_f_test}
\alias{reg_f_test}
\title{F-test taking model complexity into account
From Wikipedia: Consider two models, 1 and 2, where model 1 is 'nested' 
within model 2. Model 1 is the Restricted model, and Model 2 is the Unrestricted one. 
That is, model 1 has p1 parameters, and model 2 has p2 parameters, 
where p2 > p1, and for any choice of parameters in model 1, the same 
regression curve can be achieved by some choice of the parameters of model 2. See
https://en.wikipedia.org/wiki/F-test#Regression_problems}
\usage{
reg_f_test(bm)
}
\description{
So for our model 1 is the null model (k=0) and 2 is the sigmoidal model, then
}
\details{
F = (n-p_2) * (RSS1 - RSS2) / {RSS2 * (p_2 - p_1)}

will have an F distribution with (p_2 - p_1, n - p_2) DOF.

Here p_2 = 4 and p_1 = 2
}

