% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regulation_df}
\alias{regulation_df}
\title{Get regulation df}
\usage{
regulation_df(oui)
}
\value{
A \code{data.frame} containing the posterior traces
for all pairwise differences in regulation timing
}
\description{
Get regulation df
}
\keyword{internal}
