% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transient}
\alias{transient}
\title{Generic function to model transient expression whenever needed}
\usage{
transient(mu0, p, b, t)
}
\arguments{
\item{mu0}{Half peak expression}

\item{p}{Peak time}

\item{b}{Peak length}

\item{t}{Pseudotime}
}
\description{
Generic function to model transient expression whenever needed
}
\keyword{internal}
