% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tsigmoid}
\alias{tsigmoid}
\title{Generic function to return sigmoid whenever needed}
\usage{
tsigmoid(mu0, k, t0, t)
}
\arguments{
\item{mu0}{The 'average' expression parameter}

\item{k}{The 'activation strength' parameter}

\item{t0}{The 'activation time' parameter}

\item{t}{The latent pseudotimes}
}
\value{
A numeric vector of length \code{length(t)} that
is the sigmoid function applied to \code{t} using the parameters
\code{mu0}, \code{k} and \code{t0}.
}
\description{
Generic function to return sigmoid whenever needed
}
\examples{
tsigmoid(1, 1, 0.5, runif(10))
}
\keyword{internal}
