% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_pseudotime}
\alias{map_pseudotime}
\title{Extract the MAP pseudotime estimates from a \code{ouija_fit}}
\usage{
map_pseudotime(oui)
}
\arguments{
\item{oui}{An object of class \code{ouija_fit}.}
}
\value{
MAP pseudotime vector of length N
}
\description{
Extract the MAP pseudotime estimates from a \code{ouija_fit}
}
\examples{
data(oui)
tmap <- map_pseudotime(oui)
}
