% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\docType{data}
\name{oui}
\alias{oui}
\title{Precomputed Ouija fit}
\format{An object of class \code{ouija_fit} of length 15.}
\usage{
oui
}
\description{
An example \code{ouija_fit} to avoid
large times computing while testing.
}
\examples{
data(oui)
}
\seealso{
example_gex
}
\keyword{datasets}
