% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\name{ouija}
\alias{ouija}
\title{Fit a Ouija object.}
\usage{
ouija(x, response_type = "switch", strengths = NULL, times = NULL,
  strength_sd = NULL, time_sd = NULL, peak_times = NULL,
  bandwidths = NULL, peak_sd = NULL, bandwidth_sd = NULL,
  warn_lp = TRUE, lp_gradient_threshold = 0.01, student_df = 10,
  inference_type = c("hmc", "vb"), normalise_expression = TRUE, ...)
}
\arguments{
\item{x}{Either an \code{SCESet} from \code{scater} or a
cell-by-gene (N by G) matrix of non-negative values representing gene expression.
log2(TPM + 1) is recommended.}

\item{strengths}{G mean activation strength parameters}

\item{times}{G mean activation time parameters}

\item{strength_sd}{Optional standard deviations for k parameters}

\item{time_sd}{Optional standard deviations for t0 parameters}

\item{warn_lp}{Ouija can perform a crude check of convergence in cases where may
models are being fit and manual inspection may be cumbersome. The log-likelihood after
the burn period is regressed off the iteration number, and if the gradient of the fit
falls above a threshold (set by \code{lp_gradient_threshold}) then the user is warned.}

\item{lp_gradient_threshold}{The threshold for convergence warning. If the slope of regressing
the log-probability of the model against the iteration number falls above this value then
the user is warned.}

\item{student_df}{Degrees of freedom for the student's t likelihood}

\item{inference_type}{The type of inference to be performed, either \code{hmc} for Hamiltonian
Monte Carlo or \code{vb} for ADVI (Variational Bayes). Note that HMC is typically more accurate
but VB will be orders of magnitude faster.}

\item{normalise_expression}{Logical, default TRUE. If TRUE the data is pre-normalised
so the average peak expression is approximately 1. This makes the strength parameters
approximately comparable between genes.}

\item{...}{Additional arguments to \code{rstan::sampling}}
}
\value{
An object of type \code{ouija_fit}
}
\description{
Fit a Bayesian non-linear factor analysis model given some single-cell
gene expression data.
}
\details{
This function takes either a \code{SCESet} object or an expression matrix
and returns a \code{ouija_fit} object including posterior traces for all
variables.
}
\examples{
data(synth_gex)
oui <- ouija(synth_gex, strengths = 5 * c(1, -1, 1, -1, -1, -1), iter = 100)
}
