% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\name{ouija}
\alias{ouija}
\title{Fit a Ouija object.}
\usage{
ouija(x, response_type = "switch", switch_strengths = NULL,
  switch_times = NULL, switch_strength_sd = NULL, switch_time_sd = NULL,
  peak_times = NULL, peak_lengths = NULL, peak_time_sd = NULL,
  peak_length_sd = NULL, student_df = 10, inference_type = c("hmc", "vb"),
  normalise_expression = TRUE, ...)
}
\arguments{
\item{x}{Either an \code{SCESet} from \code{scater} or a
cell-by-gene (N by G) matrix of non-negative values representing gene expression.
log2(TPM + 1) is recommended.}

\item{response_type}{A vector declaring whether each gene exhibits "switch" or "transient"
expression. Defaults to "switch" for all genes}

\item{switch_strengths}{Prior means of switch strengths}

\item{switch_times}{Prior means of switch times}

\item{switch_strength_sd}{Prior standard deviations of switch strengths}

\item{switch_time_sd}{Prior standard deviations of switch times}

\item{peak_times}{Prior means of peak times}

\item{peak_lengths}{Prior means of peak lengths}

\item{peak_time_sd}{Prior standard deviations of peak times}

\item{peak_length_sd}{Prior standard deviations of peak lengths}

\item{student_df}{Degrees of freedom for the student's t likelihood}

\item{inference_type}{The type of inference to be performed, either \code{hmc} for Hamiltonian
Monte Carlo or \code{vb} for ADVI (Variational Bayes). Note that HMC is typically more accurate
but VB will be orders of magnitude faster.}

\item{normalise_expression}{Logical, default TRUE. If TRUE the data is pre-normalised
so the average peak expression is approximately 1. This makes the strength parameters
approximately comparable between genes.}

\item{...}{Additional arguments to \code{rstan::sampling} or \code{rstan::vb}}
}
\value{
An object of type \code{ouija_fit}
}
\description{
Fit a Bayesian non-linear factor analysis model given some single-cell
gene expression data.
}
\details{
This function takes either a \code{SCESet} object or an expression matrix
and returns a \code{ouija_fit} object including posterior traces for all
variables.
}
\examples{
data(example_gex)
response_types <- c(rep("switch", 9), rep("transient", 2))
# oui <- ouija(example_gex, response_type = response_types, iter = 100)
}
