% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\name{ouija}
\alias{ouija}
\title{Fit a Ouija object.}
\usage{
ouija(x, strengths = NULL, times = NULL, strength_sd = NULL,
  time_sd = NULL, response = c("nonlinear", "linear"), warn_lp = TRUE,
  lp_gradient_threshold = 0.01, ...)
}
\arguments{
\item{x}{Either an \code{SCESet} from \code{scater} or a
cell-by-gene (N by G) matrix of non-negative values representing gene expression.
log2(TPM + 1) is recommended.}

\item{strengths}{G mean activation strength parameters}

\item{times}{G mean activation time parameters}

\item{strength_sd}{Optional standard deviations for k parameters}

\item{time_sd}{Optional standard deviations for t0 parameters}

\item{response}{The type of factor analysis, either \code{nonlinear} (default) 
or \code{linear}}

\item{warn_lp}{Ouija can perform a crude check of convergence in cases where may
models are being fit and manual inspection may be cumbersome. The log-likelihood after
the burn period is regressed off the iteration number, and if the gradient of the fit
falls above a threshold (set by \code{lp_gradient_threshold}) then the user is warned.}

\item{lp_gradient_threshold}{The threshold for convergence warning. If the slope of regressing
the log-probability of the model against the iteration number falls above this value then
the user is warned.}

\item{...}{Additional arguments to \code{rstan::sampling}}
}
\value{
An object of type \code{ouija_fit}
}
\description{
Fit a Bayesian non-linear factor analysis model given some single-cell
gene expression data.
}
\details{
This function takes either a \code{SCESet} object or an expression matrix
and returns a \code{ouija_fit} object including posterior traces for all
variables.
}

