% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlfa.R
\name{plot_bnlfa_fit_map}
\alias{plot_bnlfa_fit_map}
\title{Plot gene expression as a function of MAP pseudotime}
\usage{
plot_bnlfa_fit_map(bm)
}
\description{
Plot gene expression as a function of MAP pseudotime
}

