% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\name{rexprs}
\alias{rexprs}
\title{Reconstructed expression}
\usage{
rexprs(oui)
}
\arguments{
\item{oui}{An object of class \code{ouija_fit}.}
}
\value{
A matrix of the same dimension as \code{oui$Y} containing 
the predicted expression.
}
\description{
The reconstructed (ie predicted) expression from the fit.
}
\examples{
data(oui)
pexp <- rexprs(oui)
}
