% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rexprs}
\alias{rexprs}
\alias{predicted_expression}
\title{Predicted expression}
\usage{
predicted_expression(oui)
}
\arguments{
\item{oui}{An object of class \code{ouija_fit}.}
}
\value{
A matrix of the same dimension as \code{oui$Y} containing 
the predicted expression.
}
\description{
The predicted (ie mean) expression from the fit.
}
\examples{
data(oui)
pexp <- rexprs(oui)
}
