% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\docType{data}
\name{synth_gex}
\alias{synth_gex}
\title{Synthetic gene expression matrix}
\format{An object of class \code{matrix} with 100 rows and 6 columns.}
\usage{
synth_gex
}
\value{
A matrix containing some synthetic gene expression data for 
100 cells and 6 genes
}
\description{
A matrix containing some synthetic gene expression data for 
100 cells and 6 genes
}
\examples{
data(synth_gex)
}
\keyword{datasets}

