% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouija.R
\docType{data}
\name{true_pst}
\alias{true_pst}
\title{Synthetic gene pseudotime vector}
\format{An object of class \code{numeric} of length 100.}
\usage{
true_pst
}
\value{
A vector with the 'true' pseudotimes for the synthetic 
gene expression data in \code{synth_gex}
}
\description{
A vector with the 'true' pseudotimes for the synthetic 
gene expression data in \code{synth_gex}
}
\examples{
data(true_pst)
}
\keyword{datasets}
